Move cache invalidation to main data store
